feat: lesson_6 0.3

+++ app/controllers/questions_controller.rb
@@ -7,6 +7,7 @@ class QuestionsController < ApplicationController
 
   def show
     question
+    @answer = @question.answers.build
   end
 
   def new
@@ -18,7 +19,7 @@ class QuestionsController < ApplicationController
   end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.build(question_params)
 
     if @question.save
       redirect_to @question, notice: 'Your question successfully created'
@@ -36,8 +37,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if question.user == current_user
+      question.destroy
+      redirect_to questions_path, notice: 'Question successfully deleted.'
+    else
+      redirect_to question, alert: 'You are not allowed to delete this question.'
+    end
   end
 
   private
